Moved source code to Source, scripts to Scripts. Moved test scripts (pytest) to Testing. The test execute, but are out of date with the code. Made top-level LICENSE file used by LicenseFrame.
