+++ Scripts/buildpyinstaller.ps1
@@ -3,4 +3,4 @@
 # mcbuzzer.py were required to make it work
 # --onefile produces a single executable
 
-pyinstaller --windowed --onefile --add-data "resources/main_banner.jpg;./resources" --add-data "resources/LICENSE.txt;./resources" --add-data "resources/ABOUT.txt;./resources" mcbuzzer.py 
+pyinstaller --windowed --onefile --add-data "resources/main_banner.jpg;./resources" --add-data "LICENSE;./resources" --add-data "resources/ABOUT.txt;./resources" Source/mcbuzzer.py
