+++ Source/competitor.py
