+++ Source/competitorGrid.py
