+++ Testing/test_conteststate.py
@@ -15,6 +15,13 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import random
+import sys
+import os
+sys.path.append(
+    os.path.join(
+        os.path.dirname(os.path.realpath(__file__)),
+        "..", "Source"))
+
 import contest
 import conteststate
 import competitor
