--- tree.py
@@ -1,33 +1,0 @@
-# MCbuzzer MATHCOUNTS Countdown Round Tool
-# Copyright (C) 2020 Park Hays
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-import math
-
-class NonPowerOfTwoCompetitorsError(Exception):
-    pass
-
-class Tree(object):
-    def __init__(self):
-        
-        self.children = []
-        self.data = None
-        
-    def initialize(self, n):
-        if math.log2(n)%1. != 0:
-            raise NonPowerOfTwoCompetitorsError()
-
-    def addChildTree(self):
-        pass
